requirements.txt
streamlit
streamlit-audiorecorder
openai-whisper
torch
soundfile
numpy
pydub
